\name{CRUDProtocol-class}
\docType{class}

\alias{class:CRUDProtocol}
\alias{CRUDProtocol-class}

\title{CRUDProtocol}
\description{
  \code{CRUDProtocol} is a base class for implementing the communication
  protocol for performing Create, Read, Update and Delete (CRUD)
  operations on a resource identified by a
  \code{\linkS4class{RestUri}}. Only HTTP is implemented by this
  package, and it should serve as useful example for other
  implementations.
}

\author{ Michael Lawrence }

\keyword{methods}
\keyword{classes}
