\name{MediaCache-class}
\docType{class}

\alias{class:MediaCache}
\alias{MediaCache-class}

\alias{$<-,MediaCache-method}
\alias{[[<-,MediaCache-method}

\alias{MediaCache}

\alias{globalRestClientCache}

\title{MediaCache}
\description{
  \code{MediaCache} is just an environment that restricts the types of
  its elements to \code{\linkS4class{Media}}. Construct an instance by
  calling \code{MediaCache}. The shared default for all REST clients is
  returned by \code{globalRestClientCache}.
}

\author{ Michael Lawrence }

\keyword{methods}
\keyword{classes}
